+++ src/components/Profile/ProfilePageMain.jsx
@@ -1,6 +1,4 @@
 import React from "react";
-import { useMutation } from "@tanstack/react-query";
-import { useNavigate } from "react-router-dom";
 import ProfileImage from "../Common/Image/ProfileImage";
 import Information from "./Information/Information";
 import Introduction from "./Introduction/Introduction";
@@ -9,20 +7,11 @@ import Skills from "./Skill/Skills";
 import Description from "./Description/Description";
 import useProfiles from "../../hooks/useProfiles";
 import Button from "../Common/Button";
-import { quit } from "../../apis/user";
-import { useSetRecoilState } from "recoil";
-import { isLoggedInState, memberIdState } from "../../recoil/atoms/auth";
-import { removeCookie } from "../../utils/cookie";
-import { newChatMemberState } from "../../recoil/atoms/newChat";
 import useModal from "../../hooks/useModal";
 import QuitModal from "./Common/QuitModal";
 
 export default function ProfilePageMain() {
-  const navigate = useNavigate();
   const [isOpen, setIsOpen] = useModal();
-  const setIsLoggedIn = useSetRecoilState(isLoggedInState);
-  const setMemberId = useSetRecoilState(memberIdState);
-  const setNewChatMembers = useSetRecoilState(newChatMemberState);
   const {
     profileQuery: {
       data: {
@@ -35,17 +24,8 @@ export default function ProfilePageMain() {
       },
       error,
     },
+    quitMutate,
   } = useProfiles();
-  const quitMutate = useMutation({
-    mutationFn: quit,
-    onSuccess: () => {
-      setIsLoggedIn(false);
-      setMemberId(null);
-      setNewChatMembers([]);
-      removeCookie("accessToken", { path: "/" });
-      navigate("/");
-    },
-  });
 
   const handleQuitButtonClick = () => {
     setIsOpen();
